Moved settings to a file
